Add an explicit dependency on rustc-std-workspace-core
When built as a proc-macro dependency with the `rustc-dep-of-std`
feature, we add an explicit `std` dependency but not one on `core`.
This means that any macros defined in `core` (such as panic-related macros)
that refer to `::core::` will fail to compile when used in
`literal-escaper`.

Work around this by adding an explicit dependency on
`rustc-std-workspace-core` when the `rustc-dep-of-std` feature is
enabled.